feat moved mcpverifier module to internal/ and added docs
