+++ internal/mcp-verifier/model.go
@@ -10,9 +10,9 @@ import (
 	"time"
 
 	"github.com/modelcontextprotocol/go-sdk/mcp"
-	"github.com/operantai/woodpecker/cmd/woodpecker-mcp-verifier/mcp-verifier/oauth"
 	"github.com/operantai/woodpecker/cmd/woodpecker-mcp-verifier/utils"
 	"github.com/operantai/woodpecker/cmd/woodpecker-mcp-verifier/vschema"
+	"github.com/operantai/woodpecker/internal/mcp-verifier/oauth"
 	"github.com/operantai/woodpecker/internal/output"
 	"github.com/spf13/viper"
 	"github.com/tmc/langchaingo/llms/openai"
@@ -90,9 +90,9 @@ func NewMCPClient(options ...Option) (IMCPClient, error) {
 	// Current module implementation of structure output is not that dynamic for the current needs where we dont
 	// know the input schema format in advance and we want to leverage it to test the tool
 	// Also you can pass the WOODPECKER_LLM_BASE_URL and should work with any OpenAI compatible APIs. You can use the
-	// OPENAI_API_KEY env var, to auth to the model. For token auth to your MCP server you can use the WOODPECKER_AUTH_HEADER env var, where you can pass your "Bearer token" and that will set the Authorization header.
+	// OPENAI_API_KEY env var, to auth the provider.
 	if mc.useAi {
-		llm, err := openai.New(openai.WithModel(viper.GetString("LLM_MODEL")), openai.WithBaseURL(viper.GetString("LLM_BASE_URL")))
+		llm, err := openai.New(openai.WithModel(viper.GetString("LLM_MODEL")), openai.WithBaseURL(viper.GetString("LLM_BASE_URL")), openai.WithToken(viper.GetString("LLM_AUTH_TOKEN")))
 		if err != nil {
 			return nil, fmt.Errorf("an error initializing the LLM client: %v", err)
 		}
